requirements.txt
streamlit==1.28.1
groq==0.4.1
python-dotenv==1.0.0
pandas==2.1.4
httpx==0.27.2
